+++ src/components/manage/Blocks/Image/Edit.jsx
@@ -265,31 +265,62 @@ class Edit extends Component {
         )}
       >
         {data.url ? (
-          <img
-            className={cx({
-              'full-width': data.align === 'full',
-              large: data.size === 'l',
-              medium: data.size === 'm',
-              small: data.size === 's',
-            })}
-            src={
-              isInternalURL(data.url)
-                ? // Backwards compat in the case that the block is storing the full server URL
-                  (() => {
-                    if (data.size === 'l')
+          <figure
+            className={cx(
+              'figure',
+              {
+                center: !Boolean(data.align),
+              },
+              data.align,
+              {
+                'full-width': data.align === 'full',
+                large: data.size === 'l',
+                medium: data.size === 'm' || !data.size,
+                small: data.size === 's',
+              },
+            )}
+          >
+            <img
+              src={
+                isInternalURL(data.url)
+                  ? // Backwards compat in the case that the block is storing the full server URL
+                    (() => {
+                      if (data.size === 'l')
+                        return `${flattenToAppURL(data.url)}/@@images/image`;
+                      if (data.size === 'm')
+                        return `${flattenToAppURL(
+                          data.url,
+                        )}/@@images/image/preview`;
+                      if (data.size === 's')
+                        return `${flattenToAppURL(
+                          data.url,
+                        )}/@@images/image/mini`;
                       return `${flattenToAppURL(data.url)}/@@images/image`;
-                    if (data.size === 'm')
-                      return `${flattenToAppURL(
-                        data.url,
-                      )}/@@images/image/preview`;
-                    if (data.size === 's')
-                      return `${flattenToAppURL(data.url)}/@@images/image/mini`;
-                    return `${flattenToAppURL(data.url)}/@@images/image`;
-                  })()
-                : data.url
-            }
-            alt={data.alt || ''}
-          />
+                    })()
+                  : data.url
+              }
+              alt={data.alt || ''}
+            />
+            {data?.description && (
+              <figcaption className="description">
+                {data?.description}
+              </figcaption>
+            )}
+            {data?.rights && (
+              <>
+                <figcaption
+                  className={cx('copyright', {
+                    nodescription: !data?.description,
+                  })}
+                >
+                  <em>Copyright: </em>
+                </figcaption>
+                <figcaption className="rights">
+                  <em>&mdash; {data?.rights}</em>
+                </figcaption>
+              </>
+            )}
+          </figure>
         ) : (
           <div>
             {this.props.editable && (
@@ -319,7 +350,16 @@ class Edit extends Component {
                               onClick={(e) => {
                                 e.stopPropagation();
                                 e.preventDefault();
-                                this.props.openObjectBrowser();
+                                this.props.openObjectBrowser({
+                                  mode: 'image',
+                                  onSelectItem: (url, description) => {
+                                    this.props.onChangeBlock(this.props.block, {
+                                      ...this.props.data,
+                                      url: url,
+                                      description: description.Description,
+                                    });
+                                  },
+                                });
                               }}
                             >
                               <Icon name={navTreeSVG} size="24px" />